refactor: move commands and events inside respective aggregates
