+++ src/application/session/aggregates/session/commands/utils.ts
@@ -1,4 +1,4 @@
-import { User, UserId } from '../session/user'
+import { User, UserId } from '../../../user'
 
 export function getUserEmailFromToken(token: string): UserId {
   return new UserId(token)
