+++ src/presentation/commands/session/chat/sendMessage.ts
@@ -1,8 +1,8 @@
 import { Server } from 'socket.io'
 import { Ack } from '../../../../application/session/message'
-import { SessionCommandHandlers } from '../../../../application/commands/sessionCommandHandlers'
 import { SessionNotifications } from '../../../notifications/sessionNotifications'
-import { SendMessageCommand } from '../../../../application/commands/chatCommands'
+import { SendMessageCommand } from '../../../../application/session/aggregates/chat/commands/chatCommands'
+import { SessionCommandHandlers } from '../../../../application/session/aggregates/session/commands/sessionCommandHandlers'
 
 /**
  * Send message command.
