+++ src/presentation/registerCommands.ts
@@ -1,6 +1,6 @@
 import { Server } from 'socket.io'
 import { connectionCommand } from './commands/connection/connection'
-import { SessionCommandHandlers } from '../application/commands/sessionCommandHandlers'
+import { SessionCommandHandlers } from '../application/session/aggregates/session/commands/sessionCommandHandlers'
 
 export async function registerCommands(io: Server) {
   connectionCommand(io, new SessionCommandHandlers())
